+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - **Breaking**: Array subset methods dependent on an `array_shape` now use the `IncompatibleArrayShapeError` error type instead of `IncompatibleDimensionalityError`
  - **Breaking**: Various array subset iterators now have validated `new` and unvalidated `new_unchecked` constructors
  - Blosc codec: disable parallel encoding/decoding for small inputs/outputs
+ - Bytes codec: optimise implementation
 
 ### Fixed
  - Bytes codec handling of complex and raw bits data types